requirements.txt
ansitable==0.11.4
cfgv==3.4.0
colored==2.3.1
contourpy==1.3.0
cycler==0.12.1
distlib==0.4.0
filelock==3.19.1
fonttools==4.60.2
identify==2.6.15
importlib_resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
nodeenv==1.9.1
numpy==1.26.4
packaging==25.0
pgraph-python==0.6.3
pillow==11.3.0
platformdirs==4.4.0
pre_commit==4.3.0
progress==1.6.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
roboticstoolbox-python==1.1.1
rtb-data==1.0.1
scipy==1.13.1
six==1.17.0
spatialgeometry==1.1.0
spatialmath-python==1.1.15
swift-sim==1.1.0
typing_extensions==4.15.0
virtualenv==20.35.4
websockets==15.0.1
zipp==3.23.0
